change stereo separation setting
First, it's not just called "stereo separation",
not "stereo separation percent", for brevity.

Second, the default is 50%, as that's a decent
all-around compromise: MODs won't have too much
stereo, and modern files will have at least
some stereo left.

For a "perfect" setup, the right choice is
to use a tmcp.ini like this:
    [TMCP]
    filter = auto
    sterep separation = 100
    [*.mod]
    filter = amiga
    stereo separation = 20