rework loadModule() error handling
Turns out that fopen() can succeed for directories,
but the resulting FILE* has super strange behavior
and is ultimately useless.
Code has been added to detect this.

This also limits the maximum size of a module file
to 64 MiB, which I hope is reasonable.

The upshot is that error handling in loadModule()
is now much more DRY.