add EBU R128 loudness analyzer
See README.md for user documentation.

What happens behind the curtains is that the module is re-opened
in a special mode that forbids playback; instead, a separate thread
runs through the entire module as fast as possible and feeds the
rendered audio data into an EBU R128 analyzer library. The computed
gain is then written into a .mod.tmcp sidecar file, where it's picked
up on playback and normalization to an adjustable target loudness
is done automagically.