initial version of bitmap font support
Using the "font" option, PC and Amiga ("Topaz") ROM bitmap fonts
can now be selected instead of the standard Inconsolata MSDF font.
There are some rough edges, and it's currently not possible to set
the font on a per-widget basis (and I'm not even sure if we really
need that anyway), but if you want the extra retro flair, be our
guest.