display effects in compact pattern visualization views
In the more size-constrained pattern visualizations,
effects used to be omitted completely. This has now
been changed: If there's unused space at the end of
the cell (or if the cell is completely empty), and
there's an effect or volume effect, this will now
be shown instead.

This makes volume slides etc. for modules with many
channels visible, adding a lot of detail to the
pattern visualization.