fake gamma correct rendering
This likely needs some explanation, so here goes ...

We're rendering straight into an sRGB framebuffer, for two reasons:
(a) that's how it has always been, and (b) that's what people are
used to when specifying colors in the config file, and how they
expect transparency to behave. It is, however, problematic in a
few areas; most noticeably, scrolling text shimmers quite a bit.
Gamma-correct rendering would fix that, and I'm in fact contemplating
on switching over to that, but for the time being, we can do the next
best thing, in true demoscene tradition: fake it!

The default settings are all bright text on dark backgrounds. We can
thus simply apply gamma on the alpha channel of text etc., and get a
pretty close facsimile of true gamma correct rendering without the
downsides. That's exactly what the new "alpha gamma" option does, and
it's preconfigured to sRGB-ish gamma. Scroll shimmering is thus
minimized a lot, but the font weight appears to be increased (which
is, in fact, completely normal, and would also happen with true
gamma-correct rendering). It's a compromise, but I think it's a good one.