reworked config loading/saving
Instead of bespoke getters and setters for each item, we now have
an enum listing the few types that exist and a method that just
gives a pointer; the rest can now be done with generic parse() and
format() functions.
This also eliminates the separate registry for enums; we now assume
that those are always 0-based and can use a simplified format for
storing the strings. (This format happens to be identical to what
ImGui::Combo() uses. What a coincidence! ;)

The registry has also been extended with min/max values for integer
and float configuration items, but those aren't used yet.