implement configuration saving / update
The approach should be minimally-invasive: If a value to write
is already specified in one of the common sections of the target
tm.ini/*.tm file, the last occurrence is replaced; even whitespace
is kept intact, if possible. If a changed value is not yet mentioned
in the INI file, a new line will be added after the last non-empty
line of the common sections.