move 'title' and 'artist' to the front and make them file-specific only
The idea here is that those two (and the hidden 'loudness' parameter)
are the ones that are most likely to change from track to track, so
we move them up to the front where they are noticed most easily.
Furthermore, we hide them from the global config UI because setting
a global artist/title override really doesn't make any sense.