make a few common key combos available while the config UI is up
ImGui loves to eat all keyboard events and there doesn't seem to be an
option to stop it from doing so and only consume keypresses it's
interested in. So we need to explicitly handle keyboard shortcuts
on an as-needed basis - currently, that's Ctrl+S, F5, and PageUp/Down,
which should cover the most common things one would do while the
config dialogs are up.