updateConfig() bugfix
When an empty sting item was replaced by a non-empty value,
the value was added as a separate line, causing the value
to stay empty and introducing a new line in the file with
bogus content.

The happened because the logic that tries to retain the
whitespace between the equals sign and the old value was
too eager and retained the newline as well.