implement automatic loop detection
Thanks to libopenmpt 0.8's new get_restart_{order,row} APIs,
we can finally have a first stab at automatic detection of
whether a module is supposed to be looped or not.

Use the new "auto loop" option to enable this functionality.
It's not super reliable, but it's usually a good starting
point. If it fails, "auto loop" can always be disabled and
the old manual "loop" option can be used on a per-module basis.